Helper function for GPDB Memory Accounting
ORCA has never fallen under the scope of GPDB's memory accounting
framework. This commit brings it one step closer by providing the
function UlSizeOfRawAlloc() to get the size of a pointer. This is needed
because GPDB will provide GPOS with a free function, however GPDB does
not have a way of determining the size of arbitrary pointers passed to
said free function. Once ORCA is tracked under memory accounting and
calls GPDB's free function with its pointers to free GPDB will in turn
call UlSizeOfRawAlloc() to ask GPOS to look into the header of the
pointer and return the size value

Signed-off-by: Kavinder Dhaliwal <kavinderd@gmail.com>